refactor(agent-toolkit): update GraphQL import paths for form tools
- Changed import paths for CreateFormTool and GetFormTool to use the new workforms.graphql file.
- Removed the old queries.graphql file references to streamline the codebase.
