+++ packages/agent-toolkit/src/core/tools/platform-api-tools/workforms-tools/get-form-tool.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod';
 import { GetFormQuery, GetFormQueryVariables } from '../../../../monday-graphql/generated/graphql';
-import { getForm } from '../../../../monday-graphql/queries.graphql';
+import { getForm } from './workforms.graphql';
 import { ToolInputType, ToolOutputType, ToolType } from '../../../tool';
 import { BaseMondayApiTool, createMondayApiAnnotations } from '../base-monday-api-tool';
 import { GraphQLDescriptions } from './workforms.consts';
